Moved slugify to uiSpec-helpers
