+++ designer/src/state/helpers/uiSpec-helpers.ts
@@ -1,6 +1,30 @@
 import {NotebookUISpec} from '../initial';
 
 /**
+ * Slugify a string, replacing special characters with less special ones
+ * @param str input string
+ * @returns url safe version of the string
+ * https://ourcodeworld.com/articles/read/255/creating-url-slugs-properly-in-javascript-including-transliteration-for-utf-8
+ */
+export const slugify = (str: string) => {
+  str = str.trim();
+  //str = str.toLowerCase();
+  // remove accents, swap ñ for n, etc
+  const from = 'ãàáäâáº½èéëêìíïîõòóöôùúüûñç·/_,:;';
+  const to = 'aaaaaeeeeeiiiiooooouuuunc------';
+  for (let i = 0, l = from.length; i < l; i++) {
+    str = str.replace(new RegExp(from.charAt(i), 'g'), to.charAt(i));
+  }
+
+  str = str
+    .replace(/[^A-Za-z0-9 -]/g, '') // remove invalid chars
+    .replace(/\s+/g, '-') // collapse whitespace and replace by -
+    .replace(/-+/g, '-'); // collapse dashes
+
+  return str;
+};
+
+/**
  * Finds the ID of the viewset (form) that contains a given view (section).
  *
  * @param state - The full notebook UI specification.
