+++ designer/src/state/uiSpec-reducer.ts
@@ -17,35 +17,11 @@ import {NotebookUISpec, FieldType, initialState} from './initial';
 import {getFieldSpec} from '../fields';
 import {ConditionType} from '../components/condition';
 import {
+  slugify,
   getViewSetForView,
   removeFieldFromSummary,
   removeFieldFromSummaryForViewset,
 } from './helpers/uiSpec-helpers';
-// eslint-disable-next-line n/no-extraneous-import
-
-/**
- * Slugify a string, replacing special characters with less special ones
- * @param str input string
- * @returns url safe version of the string
- * https://ourcodeworld.com/articles/read/255/creating-url-slugs-properly-in-javascript-including-transliteration-for-utf-8
- */
-export const slugify = (str: string) => {
-  str = str.trim();
-  //str = str.toLowerCase();
-  // remove accents, swap ñ for n, etc
-  const from = 'ãàáäâáº½èéëêìíïîõòóöôùúüûñç·/_,:;';
-  const to = 'aaaaaeeeeeiiiiooooouuuunc------';
-  for (let i = 0, l = from.length; i < l; i++) {
-    str = str.replace(new RegExp(from.charAt(i), 'g'), to.charAt(i));
-  }
-
-  str = str
-    .replace(/[^A-Za-z0-9 -]/g, '') // remove invalid chars
-    .replace(/\s+/g, '-') // collapse whitespace and replace by -
-    .replace(/-+/g, '-'); // collapse dashes
-
-  return str;
-};
 
 const uiSpecInitialState: NotebookUISpec =
   initialState.notebook['ui-specification'].present;
